Move controller into folder
