+++ lib/tokenable/controllers/tokens_controller.rb
