+++ lib/tokenable/engine.rb
@@ -1,4 +1,4 @@
-require_relative 'tokens_controller'
+require_relative 'controllers/tokens_controller'
 require_relative 'strategies/secure_password'
 
 module Tokenable
